% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.network.uncompressed.rds.R
\name{as.network.uncompressed.rds}
\alias{as.network.uncompressed.rds}
\title{Function that extracts network features}
\usage{
\method{as.network}{uncompressed.rds}(x, na.rm = FALSE, edge.check = FALSE)
}
\description{
Function that extracts network features
}

