% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraints}
\alias{constraints}
\title{Defines a function that evaluates the non-constant part of a constraint
at the provided parameter and data values}
\usage{
constraints(pars, m, M, n, kYW, kYM, kOW, kOM, DYW, DYM, DOW, DOM, K)
}
\arguments{
\item{pars}{Vector of starting values for estimating unknown parameters, vector must have names(pars) that identify the individual elements consistent with  names(work[["start"]]), where the object work is output of the get.data function}

\item{m}{Vector of sample sizes by gender, sexual identity and age group, vector must have names(m) that identify the individual elements consistent with  names(work[["m"]]), where the object work is output of the get.data function}

\item{M}{Vector of sample sizes by gender and age group, vector must have names(M) that identify the individual elements consistent with  names(work[["M"]]), where the object work is output of the get.data function}

\item{n}{4 by 3 matrix of sample mixing totals, matrix must have row.names(n) and column.names(n) that identify the individual elements consistent with  names(work[["n"]]), where the object work is output of the get.data function}

\item{kYW}{Vector of observed degree values in the sample for younger women, ordered from lowest to highest}

\item{kYM}{Vector of observed degree values in the sample for younger men, ordered from lowest to highest}

\item{kOW}{Vector of observed degree values in the sample for older women, ordered from lowest to highest}

\item{kOM}{Vector of observed degree values in the sample for older men, ordered from lowest to highest}

\item{DYW}{Vector of observed degree counts in the sample for younger women, ordered in increasing order of corresponding degree value}

\item{DYM}{Vector of observed degree counts in the sample for younger men, ordered in increasing order of corresponding degree value}

\item{DOW}{Vector of observed degree counts in the sample for older women, ordered in increasing order of corresponding degree value}

\item{DOM}{Vector of observed degree counts in the sample for older men, ordered in increasing order of corresponding degree value}

\item{K}{Degree value cutoff}
}
\value{
A vector of values corresponding to the nonconstant parts of the constraints evaluated at the provided data and parameter values
}
\description{
Defines a function that evaluates the non-constant part of a constraint
at the provided parameter and data values
}

