% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/feasRDS.rdssample.R
\name{feasRDS.rdssample}
\alias{feasRDS.rdssample}
\title{Perform Respondent Driven Sampling (RDS)}
\usage{
feasRDS.rdssample(net, nnodes = network.size(net), nsamp0, nsamp,
  replace = FALSE, coupons, seed.distribution = NULL,
  seed.composition = NULL, seed.composition.by = NULL,
  trait.variable = "disease", recruitment.rates = NULL,
  recruitment.rates.by = NULL)
}
\arguments{
\item{net}{Network to Perform RDS On}

\item{nnodes}{Number of Nodes in Network to Perform RDS Sample On}

\item{nsamp0}{Number of Seeds}

\item{nsamp}{Goal Sample Size (Is it a problem that we're approaching this as getting a certain size sample?)}

\item{replace}{Allow Replacement}

\item{coupons}{Number of Coupons Each Sampled Person Gets}

\item{seed.distribution}{Probabilities of being selected as seed if not 1/degree (default)}

\item{seed.composition}{Seed counts by variable}

\item{seed.composition.by}{Variable differentiating seed counts}

\item{trait.variable}{"Outcome" of Interest, Needs to be Binary (Diseased = 1)}

\item{recruitment.rates}{If Not Null, Variable Recruitment Rates for Number of Coupons (Matrix with # of rows = # of "recruitment.rates.by" levels,  2 columns for willingness to spread and willingness to participate named "S" and "C",  row names = levels of "recruitment.rates.by")}

\item{recruitment.rates.by}{Variable Differentiating Differential Recruitment There is Differential Recruitment}
}
\value{
RDS object
}
\description{
Perform Respondent Driven Sampling (RDS)
}

