% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/feasRDS.to.data.frame.R
\name{feasRDS.to.data.frame}
\alias{feasRDS.to.data.frame}
\title{Function that takes object produced by feasRDS.rdssample and produces a nice data frame of RDS data}
\usage{
feasRDS.to.data.frame(x, population.size = NULL)
}
\arguments{
\item{x}{Object returned by rdssample.mc}

\item{population.size}{Optional value equal to population size if supplied}
}
\value{
A data frame of RDS data
}
\description{
Function that takes object produced by feasRDS.rdssample and produces a nice data frame of RDS data
}

