% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.data.R
\name{get.data}
\alias{get.data}
\title{Function that takes data for one MSA and returns degree distributions, mixing matrices, starting values...}
\usage{
get.data(data.sub = data.sub, K = 45)
}
\arguments{
\item{data.sub}{Data for one MSA}

\item{K}{Degree cutoff to be used}
}
\value{
List of objects derived from the data used in maximum likelihood estimation and/or RDS simulation
}
\description{
Function that takes data for one MSA and returns degree distributions, mixing matrices, starting values...
}

