% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.ses.R
\name{get.ses}
\alias{get.ses}
\title{Function to get standard errors for parameters estimated by maximizing "loglik" function}
\usage{
get.ses(pars)
}
\arguments{
\item{pars}{Vector of estimates of the unknown parameters, vector must have names(pars) that identify the individual elements consistent with  names(work[["start"]]), where the object work is output of the get.data function}
}
\value{
Vector of standard error estimates for each parameter in the "pars" vector
}
\description{
Function to get standard errors for parameters estimated by maximizing "loglik" function
}

