% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/log.lik.const.1.R
\name{log.lik.const.1}
\alias{log.lik.const.1}
\title{Log Likelihood Function with Constraints Embedded}
\usage{
\method{log}{lik.const.1}(pars, m, M, n, kYW, kYM, kOW, kOM, DYW, DYM, DOW, DOM,
  K)
}
\description{
Log Likelihood Function with Constraints Embedded
}

