% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/log.lik.const.3.R
\name{log.lik.const.3}
\alias{log.lik.const.3}
\title{Log-likelihood function with constraints embedded}
\usage{
\method{log}{lik.const.3}(pars, m, M, n, kYW, kYM, kOW, kOM, DYW, DYM, DOW, DOM,
  K)
}
\description{
Log-likelihood function with constraints embedded
}

