% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reedm.R
\name{reedm}
\alias{reedm}
\title{Performs Reed-Molloy algorithm to initialize degrees}
\usage{
reedm(degrees)
}
\arguments{
\item{degrees}{Vector of desired degree counts to be matched in the simulated network, where each element is the degree count of a synthetic subject}
}
\value{
A network object with a degree distribution equal or close to the desired degree distribution
}
\description{
Performs Reed-Molloy algorithm to initialize degrees
}

